requirements.txt
yt-dlp==2025.6.9
demucs
ffmpeg-python==0.2.0
requests==2.31.0
python-dotenv==1.0.0
pyacrcloud 